Clean up, class names and package update
