+++ src/main/java/webElementsExtensions/pageWebElementsBeta/PageElementLocatorDecorator.java
@@ -1,9 +1,8 @@
-package webElementsExtensions.pageWebElements;
+package webElementsExtensions.pageWebElementsBeta;
 
 import net.sf.cglib.proxy.Enhancer;
 import net.sf.cglib.proxy.MethodInterceptor;
 import org.openqa.selenium.SearchContext;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.pagefactory.DefaultElementLocatorFactory;
 import org.openqa.selenium.support.pagefactory.DefaultFieldDecorator;
@@ -12,16 +11,23 @@ import org.openqa.selenium.support.pagefactory.FieldDecorator;
 
 import java.lang.reflect.Field;
 
-public class PageElementFieldDecorator implements FieldDecorator {
+public class PageElementLocatorDecorator implements FieldDecorator {
 
     final DefaultFieldDecorator defaultFieldDecorator;
     final SearchContext searchContext;
-//    private final WebDriver webDriver;
 
-    public PageElementFieldDecorator(SearchContext searchContext, WebDriver webDriver) {
+    public PageElementLocatorDecorator(SearchContext searchContext) {
         this.searchContext = searchContext;
-//        this.webDriver = webDriver;
-        defaultFieldDecorator = new DefaultFieldDecorator(new DefaultElementLocatorFactory(searchContext));
+        this.defaultFieldDecorator = new DefaultFieldDecorator(new DefaultElementLocatorFactory(searchContext));
+    }
+
+    @Override
+    public Object decorate(ClassLoader loader, Field field) {
+        if (PageElement.class.isAssignableFrom(field.getType()) && field.isAnnotationPresent(FindBy.class)) {
+            return getEnhancedObject(field.getType(), getElementHandler(field));
+        } else {
+            return defaultFieldDecorator.decorate(loader, field);
+        }
     }
 
     public Object getEnhancedObject(Class clazz, MethodInterceptor methodInterceptor) {
@@ -31,20 +37,8 @@ public class PageElementFieldDecorator implements FieldDecorator {
         return e.create();
     }
 
-    public Object decorate(ClassLoader loader, Field field) {
-        if (PageElement.class.isAssignableFrom(field.getType()) && field.isAnnotationPresent(FindBy.class)) {
-            return getEnhancedObject(field.getType(), getElementHandler(field));
-        } else {
-            return defaultFieldDecorator.decorate(loader, field);
-        }
-    }
-
-//    private PageElementLocator.ElementHandler getElementHandler(Field field) {
-//        return new PageElementLocator.ElementHandler(field, getLocator(field), webDriver);
-//    }
-
-    private PageElementLocator.ElementHandler getElementHandler(Field field) {
-        return new PageElementLocator.ElementHandler(field, getLocator(field));
+    private PageElementLocatorHandler getElementHandler(Field field) {
+        return new PageElementLocatorHandler(getLocator(field));
     }
 
     private ElementLocator getLocator(Field field) {
